fullstack-poubelle-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
ultralytics>=8.3.0
Pillow>=10.0.0
opencv-python>=4.10.0
torch>=2.5.0
torchvision>=0.20.0
numpy>=2.0.0,<2.3.0
pyyaml>=6.0
